adding letters

+++ backend/controllers/letterController.js
@@ -1,98 +1,59 @@
-const userService = require('../services/user');
 const letterService = require('../services/letter');
-const competitionService = require('../services/competition');
 
 exports.getLetters = async(req, res) => {
   try {
-    const letters = await letterService.getLettersByUserId(req.user.id);
+    let letters = [];
+    if (req.params.compId) {
+      letters = await letterService.getLettersByCompId(req.params.compId);
+    } else {
+      letters = await letterService.getLettersByUserId(req.user.id);
+    }
     res.json(letters);
   } catch (err) {
+    console.log(err)
     res.status(500).json({ msg: 'Server error.' });
   }
 }
 
-//invite user to join competition
-exports.addLetterToUser = async (req, res) => {
+//add letter based on letter type - see letter model
+//note comp deleted letters are handled by function in letter service
+exports.addLetter = async (req, res) => {
   try {
-    const {compName, compId, userId, startDate} = req.body
+    const {type, compId, compName, userId, userName = '', startDate = null } = req.body
 
-    letterService.addNewLetter({
-      compId, 
-      message: `You have been invited to join ${compName}!`,
-      userId, 
-      type: 'toUser',
+    let letterFields = {
+      compId,
+      message: '',
+      userId,
+      type: type,
       expireAt: startDate
-    })
+    }
 
-    res.json({ msg: 'Invite sent.' });
+    switch(type) {
+      case('toUser'):
+        letterFields.message = `${userName} invited you to join ${compName}!`
+        break;
+      case('fromUser'):
+        letterFields.message = `${userName} has requested to join this competition!`
+        break;
+      case('userAdded'):
+        letterFields.message = `${userName} approved your request to join ${compName}!`
+        break;
+      case('userKicked'):
+        letterFields.message = `You have been removed from ${compName}. The competition goal has been saved as a personal goal.`
+        break;
+      case('requestAdmin'):
+        letterFields.message = `${userName} has asked you to be an admin for ${compName}!`
+        break;
+      default:
+        return res.status(500).json({ msg: 'Invalid letter type.'}) 
+    }
+
+    const letter = await letterService.addNewLetter(letterFields)
+
+    res.json({ msg: 'Invite sent.', letter });
   } catch (err) {
-    res.status(500).json({ msg: 'Server error.' });
-  }
-}
-
-//user asked to join competition
-exports.addLetterFromUser = async (req, res) => {
-  try {
-    const {userName, compId, userId, startDate} = req.body
-    
-    letterService.addNewLetter({
-      compId, 
-      message: `${userName} has requested to join this competition!`,
-      userId, 
-      type: 'fromUser',
-      expireAt: startDate
-    })
-
-    res.json({ msg: 'Request sent.' });
-  } catch (err) {
-    res.status(500).json({ msg: 'Server error.' });
-  }
-}
-
-//tell user they were added to competition
-exports.addLetterUserAdded = async (req, res) => {
-  try {
-    const {compName, compId, userId} = req.body
-
-    letterService.addNewLetter({
-      compId, 
-      message: `Your request to join ${compName} has been approved!`,
-      userId, 
-      type: 'userAdded'
-    })
-  } catch (err) {
-    res.status(500).json({ msg: 'Server error.' });
-  }
-}
-
-//tell user they have been removed from competition
-exports.addLetterUserKicked = async (req, res) => {
-  try {
-    const {compName, compId, userId} = req.body
-
-    letterService.addNewLetter({
-      compId, 
-      message: `You have been removed from ${compName}.`,
-      userId, 
-      type: 'userKicked'
-    })
-  } catch (err) {
-    res.status(500).json({ msg: 'Server error.' });
-  }
-}
-
-//tell user competition was deleted
-exports.addLetterCompDeleted = async (req, res) => {
-  try {
-    const {compName, compId, userId} = req.body
-
-    letterService.addNewLetter({
-      compId, 
-      message: `Competition ${compName} has been deleted. You can find your progress in your goals.`,
-      userId, 
-      type: 'compDeleted'
-    })
-  } catch (err) {
+    console.log(err)
     res.status(500).json({ msg: 'Server error.' });
   }
 }
@@ -100,37 +61,18 @@ exports.addLetterCompDeleted = async (req, res) => {
 //delete letter after user dismisses it
 exports.deleteLetter = async (req, res) => {
   try {
-    const {letterId} = req.body
+    const letterId = req.params.letterId
 
-    letterService.deleteLetterById(letterId)
+    //verify letter exists
+    const letter = await letterService.getLetterById(letterId)
+    if (!letter) return res.status(404).json({msg: 'Letter not found.'})
+
+    //delete letter
+    await letterService.deleteLetterById(letterId)
+
+    res.json({ msg: 'Letter deleted.', letterId: req.params.letterId})
   } catch (err) {
-    res.status(500).json({ msg: 'Server error.' });
-  }
-}
-
-//TODO - MOVE TO USER DELETION delete all user letters when user deletes account
-exports.deleteAllUserLetters = async (req, res) => {
-  try {
-    const {userId} = req.body
-  } catch (err) {
-    res.status(500).json({ msg: 'Server error.' });
-  }
-}
-
-//TODO - MOVE TO COMPETITION DELETION delete letters when competition deleted
-exports.deleteAllCompetitionLetters = async (req, res) => {
-  try {
-
-  } catch (err) {
-    res.status(500).json({ msg: 'Server error.' });
-  }
-}
-
-//TODO - MOVE TO COMPETITION UPDATED - change date on letter when start date on competition is changed
-exports.updateLetterExpirationDate = async (req, res) => {
-  try {
-
-  } catch (err) {
+    console.log(err)
     res.status(500).json({ msg: 'Server error.' });
   }
 }
